feat: merge collection Modal
